Added persistent pool to the RNG.
Ensures that even if the HWRNG becomes broken, outputting random numbers
will still be unpredictable, if not strictly meeting their advertised
amount of entropy.

This commit won't compile as-is, (nor have I tried) as currently all the
non-volatile-storage using code is in wallet.c, along with the offsets
for where in the storage the data is kept. This needs to be refactored
into a header.